+++ sdk/core/azure-core/src/http/curl/curl_session_private.hpp
@@ -362,7 +362,7 @@ namespace Azure { namespace Core { namespace Http {
      * @param count The number of bytes to read from the network.
      * @return The actual number of bytes read from the network.
      */
-    int64_t OnRead(uint8_t* buffer, int64_t count, Azure::Core::Context const& context) override;
+    int64_t OnRead(uint8_t* buffer, size_t count, Azure::Core::Context const& context) override;
 
   public:
     /**
