+++ sdk/core/azure-core/src/io/body_stream.cpp
@@ -40,7 +40,7 @@ static_assert(sizeof(void*) >= sizeof(HANDLE), "We must be able to cast HANDLE t
 #endif
 
 // Keep reading until buffer is all fill out of the end of stream content is reached
-int64_t BodyStream::ReadToCount(uint8_t* buffer, int64_t count, Context const& context)
+int64_t BodyStream::ReadToCount(uint8_t* buffer, size_t count, Context const& context)
 {
   int64_t totalRead = 0;
 
@@ -49,7 +49,7 @@ int64_t BodyStream::ReadToCount(uint8_t* buffer, int64_t count, Context const& c
     int64_t readBytes = this->Read(buffer + totalRead, count - totalRead, context);
     totalRead += readBytes;
     // Reach all of buffer size
-    if (totalRead == count || readBytes == 0)
+    if (totalRead == static_cast<int64_t>(count) || readBytes == 0)
     {
       return totalRead;
     }
@@ -75,10 +75,11 @@ std::vector<uint8_t> BodyStream::ReadToEnd(Context const& context)
   }
 }
 
-int64_t MemoryBodyStream::OnRead(uint8_t* buffer, int64_t count, Context const& context)
+int64_t MemoryBodyStream::OnRead(uint8_t* buffer, size_t count, Context const& context)
 {
   (void)context;
-  int64_t copy_length = std::min(count, static_cast<int64_t>(this->m_length - this->m_offset));
+  int64_t copy_length = std::min(
+      static_cast<int64_t>(count), static_cast<int64_t>(this->m_length - this->m_offset));
   // Copy what's left or just the count
   std::memcpy(buffer, this->m_data + m_offset, static_cast<size_t>(copy_length));
   // move position
@@ -175,7 +176,7 @@ FileBodyStream::~FileBodyStream()
 #endif
 }
 
-int64_t FileBodyStream::OnRead(uint8_t* buffer, int64_t count, Azure::Core::Context const& context)
+int64_t FileBodyStream::OnRead(uint8_t* buffer, size_t count, Azure::Core::Context const& context)
 {
   return m_randomAccessFileBodyStream->Read(buffer, count, context);
 }
